requirements.txt
Django
python-whois
tldextract
beautifulsoup4
lxml
scikit-learn
pandas
joblib
python-magic